% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleAPP.R
\name{alleleAPP}
\alias{alleleAPP}
\title{Graphical output of alleles appeareance}
\usage{
alleleAPP(
  gene_list,
  max_alleles_bar = 8,
  file = file.path(normalizePath(tempdir()), "allele_appeareance.pdf")
)
}
\arguments{
\item{gene_list}{list of alleles appearce for each gene.}

\item{max_alleles_bar}{the number of maximum alleles per each bar plot. Defualt is 8 alleles.}

\item{file}{file path to render the plot.}
}
\value{
A bar plot for each gene with the allele appearence.
}
\description{
The \code{alleleAPP} function generates a graphical output of the alleles appeareance for each gene.
}
