% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{allelePalette}
\alias{allelePalette}
\title{Creates the allele color palette for haplotype graphical output}
\usage{
allelePalette(hap_alleles, NRA = TRUE)
}
\arguments{
\item{hap_alleles}{a list of the haplotype alleles.}
}
\value{
Haplotype allele color palette
}
\description{
\code{allelePalette} Takes a list of the haplotype alleles and returns the allele color palette.
}
