% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{draw_segment}
\alias{draw_segment}
\title{Draw lk value lines on heatmap}
\usage{
draw_segment(NR, NC, I, J, ...)
}
\arguments{
\item{NR}{number of rows.}

\item{NC}{number of columns.}

\item{I}{row index.}

\item{J}{column index.}
}
\value{
plotting lk lines.
}
\description{
\code{draw_segment} takes values for plotting text on heatmap
}
