% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoHeatmap.R
\name{genoHeatmap}
\alias{genoHeatmap}
\title{Graphical output of alleles division by genotypes}
\usage{
genoHeatmap(geno_table, chain = c("IGH", "IGK", "IGL"),
  gene_sort = "position", removeIGH = TRUE, lk_cutoff = 1,
  mark_low_lk = TRUE, html = FALSE, n_line = 4, line_length = 60,
  pseudo_genes = FALSE, ORF_genes = FALSE, file = NULL,
  color_y = NULL)
}
\arguments{
\item{geno_table}{genoytpe summary table. See details.}

\item{chain}{the IG chain: IGH,IGK,IGL. Default is IGH.}

\item{gene_sort}{if by 'name' the genes in the output are ordered lexicographically,
if by 'position' only functional genes are used and are ordered by their chromosomal location. Default is 'position'.}

\item{removeIGH}{if TRUE, 'IGH'\'IGK'\'IGL' prefix is removed from gene names.}

\item{lk_cutoff}{the lK cutoff value to be considerd low for texture layer. Defualt is lK<1.}

\item{mark_low_lk}{if TRUE, a texture is add for low lK values. Defualt is TRUE.}

\item{html}{if TRUE, an interactive html visualization is produced. Defualt is FALSE.}

\item{file}{file path for rendering the plot to pdf. If non is supplied than the plot is retured as object. Defualt is NULL.}

\item{color_y}{named list of the colors for y axis labels. Defualt is NULL.}
}
\value{
The following list is returned:
\itemize{
  \item \code{'p'}:        heat-map visualization of the genotype inference for multiple samples.
  \item \code{'width'}:    Optimal width value for rendering plot.
  \item \code{'height'}:   Optimal width value for rendering plot.
}

When a file is supplied the graph is also rendered directly to pdf.
}
\description{
The \code{genoHeatmap} function generates a graphical output of the alleles per gene in multiple samples.
}
\details{
A \code{data.frame} created by \code{inferGenotypeBaysian}.
}
