% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygousBar_html.R
\name{heterozygousBar_html}
\alias{heterozygousBar_html}
\title{Graphical output of allele heterozygousity}
\usage{
heterozygousBar_html(gene_segment, chain = c("IGH", "IGK", "IGL"))
}
\arguments{
\item{gene_segment}{a data frame of allele heterozygousity count in a given population. See details.}

\item{chain}{the IG chain: IGH,IGK,IGL. Default is IGH.}
}
\value{
An interactive stacked barplot visualization of the allele hetrouzygousity in a given population.
}
\description{
The \code{heterozygousBar_html} function generates an interactive graphical output of the alleles heterozygousity in a given population.
}
\details{
A \code{data.frame} with the following columns.
\itemize{
  \item \code{'GENE'}:     The gene call
  \item \code{'HM'}:       Count of the Individuals homozygous to the gene
  \item \code{'HT'}:       Count of the Individuals heterozygous to the gene
}
}
