% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleGenotype.R
\name{multipleGenoytpe}
\alias{multipleGenoytpe}
\title{Creates genotype graphical output division by subjects in pdf/html format}
\usage{
multipleGenoytpe(gen_table, chain = "IGH", html = FALSE,
  removeIGH = TRUE, text_size = 14, gene_sort = "position",
  facet_by = "SUBJECT", pseudo_genes = F, ORF_genes = F)
}
\arguments{
\item{gen_table}{a table of the genoytpes.}

\item{html}{character "F" or "T". "T" for html and "F" for pdf output.}
}
\value{
genotype graphical output
}
\description{
\code{multipleGenoytpe} Takes a genoytpe table and returns genotype graphical output.
}
