% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{nonReliableAllelesText_V2}
\alias{nonReliableAllelesText_V2}
\title{Creates the non reliable allele text annotation for plots}
\usage{
nonReliableAllelesText_V2(non_reliable_alleles_text, size = 3, map = F)
}
\arguments{
\item{geno_table}{a data frame of the haplotypes.}
}
\value{
Non reliable alleles text data frame for plots annotation.
}
\description{
\code{nonReliableAllelesText_V2} Takes the haplotype data frame
}
