% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{novelAlleleAnnotation_geno}
\alias{novelAlleleAnnotation_geno}
\title{Creates the novel allele text annotation for plots}
\usage{
novelAlleleAnnotation_geno(novel_allele, new_label, size = 3)
}
\arguments{
\item{novel_allele}{data frame with the novel allele cordinates.}
}
\value{
novel alleles text data frame for plots annotation.
}
\description{
\code{novelAlleleAnnotation} Takes the haplotype data frame
}
