% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{sortDFByGene}
\alias{sortDFByGene}
\title{Sort data frame by genes}
\usage{
sortDFByGene(DATA, chain = c("IGH", "IGK", "IGL"), method = c("name",
  "position"), removeIGH = FALSE, geno = FALSE, peseudo_remove = F,
  ORF_remove = F)
}
\arguments{
\item{DATA}{data frame to sort}

\item{chain}{the Ig chain: IGH,IGK,IGL. Default is IGH.}

\item{method}{the method for sorting the genes. If by 'name' the genes in the output are ordered lexicographically,
if by 'position' only functional genes are used and are ordered by their chromosomal location. Default is 'position'.}

\item{removeIGH}{if TRUE, 'IGH'\'IGK'\'IGL' prefix is removed from gene names.}
}
\value{
sorted \code{data.frame}
}
\description{
\code{sortDFByGene} Sort the \code{data.frame} by the genes names or position. For sorting by gene names the \code{sortAlleles} function by TIgGER is used.
For sorting by position the defualt package gene location list is used.
}
