% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{sortGenotype}
\alias{sortGenotype}
\title{Sort multiple genotypes}
\usage{
sortGenotype(
  geno_table,
  chain = c("IGH", "IGK", "IGL"),
  gene_sort = "position",
  removeIGH = TRUE,
  lk_cutoff = 1
)
}
\arguments{
\item{geno_table}{genoytpe summary table. See details.}

\item{chain}{the IG chain: IGH,IGK,IGL. Default is IGH.}

\item{gene_sort}{if by 'name' the genes in the output are ordered lexicographically,
if by 'position' only functional genes are used and are ordered by their chromosomal location. Default is 'position'.}

\item{removeIGH}{if TRUE, 'IGH'\'IGK'\'IGL' prefix is removed from gene names.}

\item{lk_cutoff}{the lK cutoff value to be considerd low for texture layer. Defualt is lK<1.}
}
\value{
A \code{data.frame} of the genotypes.
}
\description{
The \code{sortGenotype} function sort multiple genotype tables.
}
\details{
A \code{data.frame} created by \code{inferGenotypeBaysian}.
}
