% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{splitlines}
\alias{splitlines}
\title{Split lines of short reads assignments}
\usage{
splitlines(bottom_annot, line_width = 60)
}
\arguments{
\item{bottom_annot}{annotation text.}

\item{line_width}{the line width allowed.}
}
\value{
Seperated text to lines
}
\description{
The \code{splitlines} function sliplits the text by line width.
}
