% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjbaseVis.R
\docType{package}
\name{vdjbaseVis}
\alias{vdjbaseVis}
\alias{vdjbaseVis-package}
\title{The vdjbaseVis package}
\description{
The \code{vdjbaseVis} package provides multiple visualizations for genotypes.
}
\section{Genotype visualization}{

\code{vdjbaseVis} provides tools to infer haplotypes based on given anchor genes,
deletion detection based on relative gene usage, pooling v genes, and a single anchor gene.

\itemize{
  \item  \link{genoHeatmap}:              Genotype comparison for multiple smaples in heatmap format.
  \item  \link{multipleGenoytpe}:         Genotype visualization for single and multiple smaples.
}
}

\references{
\enumerate{
 }
}
