% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{Write_text}
\alias{Write_text}
\title{Write text annotations for heatmap graph}
\usage{
Write_text(NR, NC, I, J, ALLELE, N_ALLELES, TEXT, ...)
}
\arguments{
\item{NR}{number of rows.}

\item{NC}{number of columns.}

\item{I}{row index.}

\item{J}{column index.}

\item{ALLELE}{allele index in the individual gene box.}

\item{N_ALLELES}{number of alleles for individual in gene box.}

\item{TEXT}{annotation text.}
}
\value{
plotting text annotation.
}
\description{
\code{Write_text} takes values for plotting text on heatmap
}
