% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneUsage.R
\name{geneUsage}
\alias{geneUsage}
\title{Graphical output of gene usage in given population}
\usage{
geneUsage(
  gene_segment,
  chain = c("IGH", "IGK", "IGL"),
  plot_style = c("base", "ggplot", "plotly"),
  genePlotOrder = c("D", "J", "V")
)
}
\arguments{
\item{gene_segment}{a data frame of gene usage in a given population. See details.}

\item{chain}{the IG chain: IGH,IGK,IGL. Default is IGH.}
}
\value{
An interactive stacked barplot visualization of the allele hetrouzygousity in a given population.
}
\description{
The \code{geneUsage} function generates a graphical output of the gene usage in a given population.
}
\details{
A \code{data.frame} with the following columns.
\itemize{
  \item \code{'GENE'}:     The gene call
  \item \code{'FREQ'}:     Gene usage of each individual, comma delimited.
}
}
\examples{
 gene_segment <- data.frame(GENE = c("V1-2",'V3-3','D2-8','D3-16','J4','J6'), FREQ = rep("0.2,0.1,0.2,0.4,0.5,0.6,0.7",6))
 # plotting with base R boxplot
 p <- geneUsage(gene_segment, plot_style = "base")
 cowplot::ggdraw(p)
 # plotting with ggplot
 p <- geneUsage(gene_segment, plot_style = "ggplot")
 # plotting with plotly
 p <- geneUsage(gene_segment, plot_style = "plotly")
}
